Move mapVasuQuestion
This function is needed by generated deserialization functions
