+++ frontend/src/citizen-frontend/children/sections/vasu-and-leops/vasu/vasu-content.ts
@@ -2,13 +2,11 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import mapValues from 'lodash/mapValues'
-
 import {
   CheckboxQuestion,
   DateQuestion,
   Followup,
-  FollowupEntry,
+  mapVasuQuestion,
   MultiFieldListQuestion,
   MultiFieldQuestion,
   MultiSelectQuestion,
@@ -19,9 +17,7 @@ import {
   VasuQuestion
 } from 'lib-common/api-types/vasu'
 import { VasuContent, VasuSection } from 'lib-common/generated/api-types/vasu'
-import HelsinkiDateTime from 'lib-common/helsinki-date-time'
 import { JsonOf } from 'lib-common/json'
-import LocalDate from 'lib-common/local-date'
 
 export function isTextQuestion(
   question: VasuQuestion
@@ -79,79 +75,11 @@ export function isStaticInfoSubsection(
   return question.type === 'STATIC_INFO_SUBSECTION'
 }
 
-function isDateQuestionJson(
-  question: JsonOf<VasuQuestion>
-): question is JsonOf<DateQuestion> {
-  return question.type === 'DATE'
-}
-
-function isFollowupJson(
-  question: JsonOf<VasuQuestion>
-): question is JsonOf<Followup> {
-  return question.type === 'FOLLOWUP'
-}
-
-function isRadioGroupQuestionJson(
-  question: JsonOf<VasuQuestion>
-): question is JsonOf<RadioGroupQuestion> {
-  return question.type === 'RADIO_GROUP'
-}
-
-function isMultiSelectQuestionJson(
-  question: JsonOf<VasuQuestion>
-): question is JsonOf<MultiSelectQuestion> {
-  return question.type === 'MULTISELECT'
-}
-
 export const mapVasuContent = (content: JsonOf<VasuContent>): VasuContent => ({
   ...content,
   sections: content.sections.map((section: JsonOf<VasuSection>) => ({
     ...section,
-    questions: section.questions.map((question: JsonOf<VasuQuestion>) =>
-      isDateQuestionJson(question)
-        ? {
-            ...question,
-            value: LocalDate.parseNullableIso(question.value)
-          }
-        : isFollowupJson(question)
-          ? {
-              ...question,
-              value: question.value.map((entry: JsonOf<FollowupEntry>) => ({
-                ...entry,
-                date: LocalDate.parseIso(entry.date),
-                edited: entry.edited && {
-                  ...entry.edited,
-                  editedAt: LocalDate.parseIso(entry.edited.editedAt)
-                },
-                createdDate:
-                  typeof entry.createdDate === 'string'
-                    ? HelsinkiDateTime.parseIso(entry.createdDate)
-                    : undefined
-              }))
-            }
-          : isRadioGroupQuestionJson(question)
-            ? {
-                ...question,
-                dateRange: question.dateRange && {
-                  start: LocalDate.parseIso(question.dateRange.start),
-                  end: LocalDate.parseIso(question.dateRange.end)
-                }
-              }
-            : isMultiSelectQuestionJson(question)
-              ? {
-                  ...question,
-                  dateValue:
-                    question.dateValue &&
-                    mapValues(question.dateValue, (v) => LocalDate.parseIso(v)),
-                  dateRangeValue:
-                    question.dateRangeValue &&
-                    mapValues(question.dateRangeValue, (v) => ({
-                      start: LocalDate.parseIso(v.start),
-                      end: LocalDate.parseIso(v.end)
-                    }))
-                }
-              : question
-    )
+    questions: section.questions.map(mapVasuQuestion)
   }))
 })
